feat(fintual): collapse ingestion into the FintualPerformance service (#293)
