+++ src/fintual/fold.ts
